Combined commit: ART-side implementation of issue #883.
Comments from individual commits below.

Add ToyRawInput to test RawInputSource

Improve some error messages

Local refactoring

Local refactoring

Fix configuration of IntProducer

Clean up layout

Improve format of error message

Improve formatting of error message

Partial implementation and test of reconstitutes

Remove RawInputSource::makeEvent

Cleanup formatting

Checkpoing: renaming InputSource

Checkpoint after rebase

Checkpoint: First InputSource as pure interface

Fix typo in installed header name

Move registerProducts out of InputSource interface

Add default implementation for readEvent and skipEvents to InputSource.

Remove doEndRun and doEndSubRun from InputSource.

Remove calls to doEndRun and doEndSubRun

Rename repeat() to repeat_() and make it private.

Remove call to InputSource::repeat()

Fix bad layout.

Complete reworking of public interface.

Checkpoint

First skeleton of FileReaderSource

Checkpoint: looking for Brian or Rob

Checkpoint before renaming EventCorruption

Change all references to EventCorruption to DataCorruption instead.

Checkpoint

Replace all uses of boost::shared_ptr using do_nothing_policy with exempt_ptr.

"We're not fixing it, we're just using it." Sorry, I had to fix this one:
 I have a sensitive stomach.

Not a whole lot of progress, but code compiles and (old) tests pass.

Sort sources for ease of finding things.

Implementation of makeRunPrincipal; two more to go.

Implement remaining functions in PrincipalMaker.

Part 1 of (probably) 2: abstract functionality out of DataViewImpl so it can be called from places that don't inherit from it.

Guard CLHEP use to prevent nullptr troubles.

Handle all the work of putting a product in the event. Lots of stuff still to do here.

Revert "Guard CLHEP use to prevent nullptr troubles."

This reverts commit 09c647b9f5ee1af6519fa41be063a1c4ee457c57.

After testing, we don't have a reliable solution for
avoiding redfining the nullptr macro, so we have to just let it
happen.

Checkpoint before splitting ProductPrincipalPutter.

Split ProductPrincipalPutter into two new headers per discussion with MP.

Checkpoint.

First simple test of ToyRawInput: let the debugging commence!

Improve interaction with ParameterSet and indent.

std::basic_string<char> -> std::string to attempt to trace product problem.

Fix pastos in EventPrincipal and SubRunPrincipal.

Move source signal emission from DRISI and friends into EventPrincipal; retire source construction signals.

Need empty ParameterSet in module description.

Add missing implementation of readFileBlock().

Runs still start at 1 even though the invalid number is still -1.

a bare string product appears to be problematic -- more investigation in a different forum.

Test for bare string as product for further investigation: currently fails.

Checkpoint

Checkpoint.

Fix out-of-date usage message.

Fix behavior of FileReaderSource with respect to moving to the next file.

Round out the tests and put some more sanity checks in FileReaderSource.

Fix erroneous conflict resolution.

Reorder checks.

Add missing return.

Fix remaining fixable compilation warnings.

+++ test/Integration/SimpleDerivedAnalyzer_module.cc
@@ -72,7 +72,7 @@ void verify_elements(std::vector<T> const& ptrs,
 {
   for (size_t k = 0; k != sz; ++k)
     {
-      assert(ptrs[k]->key == sz-k+event_num);
+      assert((unsigned)ptrs[k]->key == sz-k+event_num);
       double expect = 1.5 * k + 100.0;
       assert(ptrs[k]->value == expect);
       assert(ptrs[k]->dummy() == 16.25);
@@ -194,7 +194,7 @@ SimpleDerivedAnalyzer::test_PtrVector( art::Event const & e ) const
     for( product_t::const_iterator b = h->begin()
            , e = h->end(); b!= e; ++b ) {
       int k = b - h->begin();
-      if( (*b)->key != sz-k+event_num ) {
+      if( (unsigned)(*b)->key != sz-k+event_num ) {
         std::cerr
           << "ValueMismatch at position " << k
           << " expected key " << sz-k+event_num
